requirements.txt
# Combined requirements for unified backend
# Run: pip install -r requirements.txt

# Discord bot
discord.py>=2.3.0
pytz>=2023.3

# Cohort scheduling algorithm
cohort-scheduler @ git+https://github.com/cpdally/cohort-scheduler.git

# FastAPI web server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
anthropic>=0.40.0
pyjwt>=2.8.0
python-multipart>=0.0.6  # For file uploads

# Shared
python-dotenv>=1.0.0

# Database (SQLAlchemy Core + Alembic)
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Notifications
sendgrid>=6.11.0
apscheduler>=3.10.0
PyYAML>=6.0

# Google Calendar API
google-api-python-client>=2.100.0
google-auth>=2.25.0

# Error tracking
sentry-sdk[fastapi]>=1.40.0

# Development
ruff>=0.8.0
